Update gemspec to include sharepoint-http-auth
Not being much beyond a Ruby beginner compared to other programming languages, I *think* that the sharepoint-http-auth.rb file needs to be included in this gemspec so that when I 'gem build' and 'gem install', I no longer get 'unable to load' errors. In case it matters, I'm using rbenv.

Thanks for your efforts though. This looks great compared to the Perl ?mess? that is/was to be my fallback.